WIP: Creating plugins module
