+++ goodload-plugins/goodload-sqlite-sink/src/main/java/org/goodload/goodload/plugin/datasink/sqlite/models/StepSkeletonEntity.java
@@ -14,14 +14,13 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.goodload.goodload.reporting.datasink.sqlite.models;
+package org.goodload.goodload.plugin.datasink.sqlite.models;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.annotations.Where;
 
 import java.io.Serial;
 import java.io.Serializable;
@@ -33,23 +32,29 @@ import java.util.List;
  * @since 1.0
  */
 @Entity
-@Table(name = "simulation_metadata")
+@Table(name = "action_metadata")
+@NoArgsConstructor
 @Getter
 @Setter
-@NoArgsConstructor
-public class SimulationEntity implements Serializable {
+public class StepSkeletonEntity implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
     @Id
-    private String simulationId;
+    private String stepId;
 
     @Column(nullable = false)
     @NotNull
-    private String simulationName;
+    private String stepName;
+
+    @Column(nullable = false)
+    @NotNull
+    private String simulationId;
+
+    @Column
+    private String parentStepId;
 
     @OneToMany(cascade = CascadeType.ALL)
-    @JoinColumn(name = "simulationId")
-    @Where(clause = "parent_step_id IS NULL")
-    private List<StepSkeletonEntity> steps;
+    @JoinColumn(name = "parentStepId")
+    private List<StepSkeletonEntity> subSteps;
 }
