+++ goodload-plugins/goodload-sqlite-sink/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+org.goodload.goodload.plugin.datasink.sqlite.SQLiteSinkAutoConfiguration
